# Perf tuning for gcc + aarch64